pyproject.toml
[tool.poetry]
name = "qshed-scheduler"
version = "0.1.0"
description = ""
authors = ["nixoninnes"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
APScheduler = "^3.9.1"
qshed-client = {path = "packages/qshed_client-0.1.0-py3-none-any.whl"}
SQLAlchemy = "^1.4.36"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
